chore(client): moved config to interfaces
