+++ src/index.ts
@@ -1,8 +1,8 @@
 export * from './AppleMusicAPI';
-export * from './clientConfiguration';
 
 export * from './appleMusicError';
 export * from './util/CalendarDate';
 
 // Interfaces
+export * from './interfaces/Config';
 export * from './interfaces/AppleMusic';
